requirements.txt
Mezzanine==3.1.10
South==1.0.2
python-memcached==1.54
django-contact-form==1.0
djangorestframework==3.1.0
geocoder==1.3.1
geopy==1.10.0
